+++ example/lib/flutter_ffmpeg_api_wrapper.dart
@@ -28,11 +28,11 @@ final FlutterFFmpegConfig _flutterFFmpegConfig = new FlutterFFmpegConfig();
 final FlutterFFmpeg _flutterFFmpeg = new FlutterFFmpeg();
 final FlutterFFprobe _flutterFFprobe = new FlutterFFprobe();
 
-void enableLogCallback(LogCallback callback) {
+void enableLogCallback(LogCallback? callback) {
   _flutterFFmpegConfig.enableLogCallback(callback);
 }
 
-void enableStatisticsCallback(StatisticsCallback callback) {
+void enableStatisticsCallback(StatisticsCallback? callback) {
   _flutterFFmpegConfig.enableStatisticsCallback(callback);
 }
 
@@ -143,6 +143,10 @@ Future<String> registerNewFFmpegPipe() async {
   return await _flutterFFmpegConfig.registerNewFFmpegPipe();
 }
 
+Future<void> closeFFmpegPipe(String ffmpegPipePath) async {
+  return await _flutterFFmpegConfig.closeFFmpegPipe(ffmpegPipePath);
+}
+
 Future<void> setEnvironmentVariable(
     String variableName, String variableValue) async {
   return await _flutterFFmpegConfig.setEnvironmentVariable(
@@ -153,6 +157,6 @@ Future<List<FFmpegExecution>> listFFmpegExecutions() async {
   return await _flutterFFmpeg.listExecutions();
 }
 
-List<String> parseArguments(command) {
+List<String>? parseArguments(command) {
   return FlutterFFmpeg.parseArguments(command);
 }